requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
google-api-python-client>=2.90.0
google-auth>=2.20.0
google-auth-httplib2>=0.1.0
requests>=2.30.0
python-dotenv>=1.0
openai>=0.27.0
pydantic>=1.10
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0
streamlit>=1.25.0
plotly>=5.15.0
dateparser>=1.1.1
statsmodels>=0.14.0
folium>=0.14.0
streamlit-folium>=0.11.0
redis>=4.5.0
